requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
requests==2.31.0  # <--- Make sure this line exists
python-dotenv==1.0.0
gunicorn==21.2.0